Calculate the mixture response from one of three different approaches: IA, GCA, or Hazard Quotient

Calculate the combined response of multiple chemicals. It calculates the generalized concentration addition response, the independent action response, and a hazard quotient

Usage

calc_concentration_response(
  C_invitro,
  hill_params,
  max_mult = 1.5,
  fixed = FALSE
)

Arguments

C_invitro

in vitro concentrations

hill_params

output from fit_hill()

max_mult

upper bound multiplier for max response

fixed

if TRUE, sd = 0

Value

list of data frames

Examples

C_invitro <- list(
  matrix(1:8 / 1e3, ncol = 2, dimnames = list(NULL, c("c1", "c2"))),
  matrix(9:16 / 1e3, ncol = 2, dimnames = list(NULL, c("c1", "c2")))
)
hill_params <- fit_hill(
  data.frame(chem = rep(c("c1", "c2"), each = 3),
             logc = c(-1, 0, 1, 0, 1, 2),
             resp = c(10, 5, 0, 4, 2, 0) / 10),
  chem = "chem"
)

calc_concentration_response(C_invitro, hill_params)
#> [[1]]
#>   sample      GCA.Eff       IA.Eff  GCA.HQ.10   IA.HQ.10
#> 1      1 0.0076864671 0.0077364934 0.55601871 0.55600659
#> 2      2 0.0006767173 0.0006770262 0.01534725 0.01534744
#> 3      3 0.0722956637 0.0741979092 1.45744108 1.45745710
#> 4      4 0.0273376842 0.0274220130 0.39699362 0.39699367
#> 
#> [[2]]
#>   sample    GCA.Eff     IA.Eff GCA.HQ.10  IA.HQ.10
#> 1      1 0.18688743 0.18693404 3.7723954 3.7724269
#> 2      2 0.01226720 0.01227572 0.3474801 0.3474778
#> 3      3 0.01883785 0.01894890 1.7068482 1.7068523
#> 4      4 0.12791787 0.12799417 2.2230433 2.2230437
#> 
calc_concentration_response(C_invitro, hill_params, fixed = TRUE)
#> [[1]]
#>   sample   GCA.Eff    IA.Eff GCA.HQ.10 IA.HQ.10
#> 1      1 0.2393398 0.2742427  13.49985     13.5
#> 2      2 0.3102525 0.3493325  23.39997     23.4
#> 3      3 0.3479706 0.3866123  33.29939     33.3
#> 4      4 0.3713827 0.4087487  43.19904     43.2
#> 
#> [[2]]
#>   sample   GCA.Eff    IA.Eff GCA.HQ.10 IA.HQ.10
#> 1      1 0.4200784 0.4517192  92.70122     92.7
#> 2      2 0.4246337 0.4554551 102.60133    102.6
#> 3      3 0.4284513 0.4585306 112.50143    112.5
#> 4      4 0.4316926 0.4611019 122.40154    122.4
#>