Get response quantiles

Get response quantiles

Usage

resp_quantiles(
  resp,
  metric = c("GCA.Eff", "IA.Eff", "GCA.HQ.10", "IA.HQ.10"),
  assays = NULL,
  assay_summary = FALSE,
  assay_quantiles = c(Median = 0.5),
  summary_quantiles = c(`10th percentile` = 0.1)
)

Arguments

resp

calculated mixture response output from calculate_response.

metric

response metric, one of "GCA.Eff", "IA.Eff", "GCA.HQ.10" or "IA.HQ.10".

assays

assays to summarize. If NULL and multiple assays exist, then the first assay is used.

assay_summary

boolean indicating whether to summarize across assays.

assay_quantiles

numeric vector of assay quantiles.

summary_quantiles

numeric vector of quantiles to compute across all assay quantiles.

Value

data frame with computed response quantiles.

Details

The columns of the returned data frame will vary based on the inputs. If assays is specified and assay_summary is FALSE, then the resulting data frame will have an assay column. If assay_summary is TRUE, then the data frame will have an summary_quantile column.