Fit 2- or 3-parameter Hill model

Fit 2- or 3-parameter Hill model

Usage

fit_hill(x, conc = "logc", resp = "resp", fixed_slope = TRUE)

Arguments

x

data frame or list of data frames.

conc

column name of base-10 log scaled concentration.

resp

column name of response.

fixed_slope

if TRUE, slope is fixed at 1.

Value

Fit parameters and other stats. "tp" is the top asymptote and "logAC50" is the 50% response concentration. If the computation of the standard deviations of these two parameters fails, then the standard deviation is set equal to the parameter estimate and is indicated by the respective imputed flag being TRUE.

Examples

# Single chemical
data <- geo_tox_data$dose_response[
  geo_tox_data$dose_response$casn == "1582-09-8",
]
fit_hill(data)
#>         tp   tp.sd  logAC50 logAC50.sd slope slope.sd  logc_min logc_max
#> 1 5.664257 1.38922 1.832212  0.1320382     1        0 -1.522879        2
#>     resp_min resp_max      AIC tp.sd.imputed logAC50.sd.imputed
#> 1 -0.2746569 4.720214 38.09679         FALSE              FALSE

# Multiple chemicals
data <- split(geo_tox_data$dose_response, ~casn)
# 2-param
fit_hill(data)
#>        name       tp     tp.sd  logAC50 logAC50.sd slope slope.sd  logc_min
#> 1 1582-09-8 5.664257 1.3892202 1.832212  0.1320382     1        0 -1.522879
#> 2   51-28-5 4.351953 4.3519525 1.778738  0.2015090     1        0 -1.522879
#> 3  510-15-6 8.757236 8.2408031 1.978500  0.3466533     1        0 -1.522879
#> 4   72-43-5 3.737664 0.7102053 1.532911  0.1967713     1        0 -1.522879
#> 5   76-44-8 4.515815 0.3513598 1.381169  0.1004316     1        0 -1.522879
#> 6   77-47-4 1.471197 0.4960156 1.782049  0.2971874     1        0 -1.522879
#> 7   87-86-5 3.670549 0.3538687 1.087010  0.1278682     1        0 -1.698970
#> 8   95-95-4 3.331832 3.3318324 1.966358  0.2625026     1        0 -1.522879
#>   logc_max   resp_min resp_max      AIC tp.sd.imputed logAC50.sd.imputed
#> 1  2.00000 -0.2746569 4.720214 38.09679         FALSE              FALSE
#> 2  2.00000 -1.4737999 3.626627 39.76412          TRUE              FALSE
#> 3  2.00000 -0.5048061 7.297697 42.70263         FALSE              FALSE
#> 4  2.00000 -0.7303619 3.114720 34.95177         FALSE              FALSE
#> 5  2.00000 -0.9074337 3.843778 36.13351         FALSE              FALSE
#> 6  2.00000 -0.5950514 1.225998 24.64278         FALSE              FALSE
#> 7  1.69897 -0.6636353 3.058791 13.17187         FALSE              FALSE
#> 8  2.00000 -0.6495255 2.776527 34.11884          TRUE              FALSE
# 3-param
fit_hill(data, fixed_slope = FALSE)
#>        name       tp     tp.sd   logAC50 logAC50.sd    slope   slope.sd
#> 1 1582-09-8 5.664257 1.0896069 1.7390364 0.12851792 1.632874  0.3132721
#> 2   51-28-5 4.351953 3.2757664 1.7738204 0.39828065 2.976161  3.0652041
#> 3  510-15-6 8.757236 3.0112456 1.7319108 0.20853276 2.191570  0.9838474
#> 4   72-43-5 3.048893 0.2056023 1.3380156 0.08472913 3.810866  1.4801096
#> 5   76-44-8 3.811537 0.3001249 1.2100606 0.08623949 1.493623  0.3156792
#> 6   77-47-4 1.471197 1.1421995 1.7801156 0.57221708 1.424830  0.9856094
#> 7   87-86-5 3.240787 0.4285640 0.9554403 0.15184890 1.252028  0.2923098
#> 8   95-95-4 3.331832 8.7820695 1.9324691 0.57810113 8.000000 12.5456495
#>    logc_min logc_max   resp_min resp_max      AIC tp.sd.imputed
#> 1 -1.522879  2.00000 -0.2746569 4.720214 17.61138         FALSE
#> 2 -1.522879  2.00000 -1.4737999 3.626627 23.94616         FALSE
#> 3 -1.522879  2.00000 -0.5048061 7.297697 16.63602         FALSE
#> 4 -1.522879  2.00000 -0.7303619 3.114720 21.89020         FALSE
#> 5 -1.522879  2.00000 -0.9074337 3.843778 33.73297         FALSE
#> 6 -1.522879  2.00000 -0.5950514 1.225998 24.45038         FALSE
#> 7 -1.698970  1.69897 -0.6636353 3.058791 14.05871         FALSE
#> 8 -1.522879  2.00000 -0.6495255 2.776527 12.91483         FALSE
#>   logAC50.sd.imputed
#> 1              FALSE
#> 2              FALSE
#> 3              FALSE
#> 4              FALSE
#> 5              FALSE
#> 6              FALSE
#> 7              FALSE
#> 8              FALSE