Simulate external exposure

Simulate external exposure

Usage

simulate_exposure(
  x,
  expos_mean = "mean",
  expos_sd = "sd",
  expos_label = "casn",
  n = 1000
)

Arguments

x

data frame or list of data frames containing exposure data.

expos_mean

column name of mean values.

expos_sd

column name of standard deviations.

expos_label

column name of labeling term, required if x has more than one row.

n

simulated sample size.

Value

list of matrices containing inhalation rates. Matrix columns are named using the values in the expos_label column for more than one data frame row. Columns are sorted to have consistent order across functions.

Examples

# Single data frame
x <- data.frame(mean = 1:3, sd = (1:3) / 10, casn = letters[1:3])
simulate_exposure(x, n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 1.0390873 1.992394 2.721191
#> [2,] 1.1053655 2.352546 3.462302
#> [3,] 0.9444516 2.122926 3.053786
#> [4,] 0.9370587 1.667883 2.596518
#> [5,] 1.0303365 1.790308 2.775329
#> 

# List of 2 data frames
y <- data.frame(mean = 4:6, sd = 0.1, casn = letters[1:3])
simulate_exposure(list(loc1 = x, loc2 = y), n = 5)
#> $loc1
#>              a        b        c
#> [1,] 0.9010575 2.065042 3.028065
#> [2,] 0.9152033 1.835347 2.843417
#> [3,] 0.9089704 1.981799 2.847017
#> [4,] 1.0409962 2.098811 3.173462
#> [5,] 0.7179948 2.208323 2.379852
#> 
#> $loc2
#>             a        b        c
#> [1,] 4.053089 4.968356 5.891600
#> [2,] 3.820255 5.166885 5.976132
#> [3,] 3.967550 4.985696 5.804117
#> [4,] 3.997974 5.159575 6.061909
#> [5,] 4.049643 4.988856 6.033413
#> 

# Input has custom column names
z <- data.frame(ave = 1:3, stdev = (1:3) / 10, chnm = letters[1:3])
simulate_exposure(z,
                  expos_mean = "ave",
                  expos_sd = "stdev",
                  expos_label = "chnm",
                  n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 0.8957941 1.832338 3.175269
#> [2,] 1.1039255 2.079109 2.628338
#> [3,] 0.9334389 2.048909 2.732970
#> [4,] 1.0803473 2.523459 2.546992
#> [5,] 0.8482680 1.852425 2.824445
#>