Simulate external exposure

Simulate external exposure

Usage

simulate_exposure(
  x,
  expos_mean = "mean",
  expos_sd = "sd",
  expos_label = "casn",
  n = 1000
)

Arguments

x

data frame or list of data frames containing exposure data.

expos_mean

column name of mean values.

expos_sd

column name of standard deviations.

expos_label

column name of labeling term, required if x has more than one row.

n

simulated sample size.

Value

list of matrices containing inhalation rates. Matrix columns are named using the values in the expos_label column for more than one data frame row. Columns are sorted to have consistent order across functions.

Examples

# Single data frame
x <- data.frame(mean = 1:3, sd = (1:3) / 10, casn = letters[1:3])
simulate_exposure(x, n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 0.9652912 2.001517 3.280278
#> [2,] 0.9901449 2.186169 3.552095
#> [3,] 1.0034766 1.863050 2.788554
#> [4,] 1.0386127 2.067480 3.002553
#> [5,] 1.0020831 1.917572 3.610257
#> 

# List of 2 data frames
y <- data.frame(mean = 4:6, sd = 0.1, casn = letters[1:3])
simulate_exposure(list(loc1 = x, loc2 = y), n = 5)
#> $loc1
#>              a        b        c
#> [1,] 0.8658314 1.830950 3.189070
#> [2,] 1.1158979 1.807686 3.630376
#> [3,] 0.9796791 2.203498 2.815879
#> [4,] 0.9621971 1.700789 2.509609
#> [5,] 1.1736111 1.763036 2.996868
#> 
#> $loc2
#>             a        b        c
#> [1,] 3.934349 5.051433 5.982218
#> [2,] 3.933047 4.824862 6.074097
#> [3,] 3.952141 5.089360 5.900256
#> [4,] 4.131946 5.022304 5.706102
#> [5,] 4.063656 5.058082 6.071902
#> 

# Input has custom column names
z <- data.frame(ave = 1:3, stdev = (1:3) / 10, chnm = letters[1:3])
simulate_exposure(z,
                  expos_mean = "ave",
                  expos_sd = "stdev",
                  expos_label = "chnm",
                  n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 0.9301995 1.829857 2.650184
#> [2,] 0.8105874 2.113240 2.856099
#> [3,] 1.0076299 2.230442 3.034604
#> [4,] 1.0875309 1.848761 2.469585
#> [5,] 1.0453827 1.902148 2.577708
#>