Simulate external exposure

Simulate external exposure

Usage

simulate_exposure(x, mean = "mean", sd = "sd", label = "casn", n = 1000)

Arguments

x

data frame or list of data frames.

mean

column name of mean values.

sd

column name of standard deviations.

label

column name of labeling term, required if x has more than one row.

n

simulated sample size.

Value

A matrix or list of matrices containing inhalation rates. Matrix columns are named using the values in the label column for more than one data frame row.

Examples

data <- split(geo_tox_data$exposure, ~FIPS)

# Single data frame
simulate_exposure(data[[1]], n = 5)
#> [1] 0.000000e+00 6.594557e-08 2.562170e-08 0.000000e+00 5.082199e-08
#> [6] 7.808398e-08 0.000000e+00 6.263525e-08

# List of data frames
simulate_exposure(data[1:3], n = 5)
#> $`37001`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 9.004350e-05 1.996969e-08        0       0       0 3.864607e-10
#> [2,] 8.064525e-05 4.736435e-08        0       0       0 3.188015e-10
#> [3,] 9.044574e-05 3.416418e-08        0       0       0 1.756318e-10
#> [4,] 9.378640e-05 9.352764e-08        0       0       0 4.638995e-10
#> [5,] 1.202382e-04 2.483928e-08        0       0       0 8.990072e-11
#>           87-86-5 95-95-4
#> [1,] 7.558460e-09       0
#> [2,] 8.534327e-09       0
#> [3,] 8.555065e-09       0
#> [4,] 1.227381e-08       0
#> [5,] 9.708806e-09       0
#> 
#> $`37003`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 1.617462e-04 1.274182e-08        0       0       0 1.119285e-09
#> [2,] 1.800113e-04 1.566815e-08        0       0       0 8.903442e-10
#> [3,] 9.320815e-05 1.041361e-08        0       0       0 8.240105e-10
#> [4,] 1.633125e-04 1.206246e-08        0       0       0 1.181102e-09
#> [5,] 1.761066e-04 1.123618e-08        0       0       0 8.567265e-10
#>           87-86-5      95-95-4
#> [1,] 1.081675e-07 1.817795e-13
#> [2,] 8.589971e-08 3.404111e-13
#> [3,] 7.802145e-08 4.515660e-13
#> [4,] 1.057492e-07 1.107349e-12
#> [5,] 8.601744e-08 7.030613e-13
#> 
#> $`37005`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 7.587793e-05 2.241753e-09        0       0       0 1.499447e-10
#> [2,] 7.223240e-05 2.973578e-09        0       0       0 2.056294e-10
#> [3,] 5.812482e-05 1.807864e-09        0       0       0 1.787431e-10
#> [4,] 5.607397e-05 1.524453e-09        0       0       0 2.219957e-10
#> [5,] 6.559516e-05 1.600162e-09        0       0       0 2.987922e-10
#>           87-86-5      95-95-4
#> [1,] 3.423179e-09 6.496239e-11
#> [2,] 4.176794e-09 4.405243e-11
#> [3,] 3.146204e-09 6.097588e-11
#> [4,] 4.425310e-09 3.702928e-11
#> [5,] 3.154506e-09 1.679080e-11
#>