Simulate external exposure

Simulate external exposure

Usage

simulate_exposure(
  x,
  expos_mean = "mean",
  expos_sd = "sd",
  expos_label = "casn",
  n = 1000
)

Arguments

x

data frame or list of data frames.

expos_mean

column name of mean values.

expos_sd

column name of standard deviations.

expos_label

column name of labeling term, required if x has more than one row.

n

simulated sample size.

Value

A list of matrices containing inhalation rates. Matrix columns are named using the values in the expos_label column for more than one data frame row.

Examples

data <- split(geo_tox_data$exposure, ~FIPS)

# Single data frame
simulate_exposure(data[[1]], n = 5)
#> [[1]]
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 1.090811e-04 5.848432e-08        0       0       0 1.056507e-10
#> [2,] 8.750059e-05 6.770802e-08        0       0       0 4.588745e-10
#> [3,] 1.162021e-04 9.743651e-09        0       0       0 5.598812e-10
#> [4,] 1.243120e-04 7.274260e-08        0       0       0 2.099088e-10
#> [5,] 8.203307e-05 3.381474e-08        0       0       0 5.183190e-10
#>           87-86-5 95-95-4
#> [1,] 1.323982e-08       0
#> [2,] 1.284609e-08       0
#> [3,] 1.577227e-08       0
#> [4,] 6.605943e-09       0
#> [5,] 1.091679e-08       0
#> 

# List of data frames
simulate_exposure(data[1:3], n = 5)
#> $`37001`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 1.080396e-04 3.457350e-08        0       0       0 2.508143e-10
#> [2,] 9.949791e-05 4.693337e-08        0       0       0 2.757947e-10
#> [3,] 7.429448e-05 6.559605e-08        0       0       0 3.560158e-10
#> [4,] 9.147674e-05 4.087990e-08        0       0       0 2.922935e-10
#> [5,] 9.151345e-05 1.258854e-08        0       0       0 4.673050e-10
#>           87-86-5 95-95-4
#> [1,] 1.315848e-08       0
#> [2,] 7.914796e-09       0
#> [3,] 9.840136e-09       0
#> [4,] 1.291400e-08       0
#> [5,] 1.042789e-08       0
#> 
#> $`37003`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 1.998010e-04 1.272200e-08        0       0       0 1.000412e-09
#> [2,] 1.686504e-04 1.279909e-08        0       0       0 1.038144e-09
#> [3,] 1.763445e-04 1.170159e-08        0       0       0 1.055774e-09
#> [4,] 9.485285e-05 1.056577e-08        0       0       0 8.827892e-10
#> [5,] 1.367047e-04 1.257312e-08        0       0       0 1.102671e-09
#>           87-86-5      95-95-4
#> [1,] 1.069074e-07 3.241957e-13
#> [2,] 1.573573e-07 9.762800e-13
#> [3,] 1.104208e-07 4.969955e-13
#> [4,] 7.295532e-08 1.661638e-13
#> [5,] 1.431837e-07 6.337565e-13
#> 
#> $`37005`
#>         1582-09-8      51-28-5 510-15-6 72-43-5 76-44-8      77-47-4
#> [1,] 6.959053e-05 2.559101e-09        0       0       0 1.975041e-10
#> [2,] 6.891846e-05 1.795468e-09        0       0       0 3.133554e-10
#> [3,] 8.032106e-05 2.366660e-09        0       0       0 2.550301e-10
#> [4,] 7.094195e-05 1.724309e-09        0       0       0 2.296631e-10
#> [5,] 5.500838e-05 2.808881e-09        0       0       0 2.439517e-10
#>           87-86-5      95-95-4
#> [1,] 4.477279e-09 4.673951e-11
#> [2,] 3.948739e-09 3.487048e-11
#> [3,] 3.939988e-09 6.348754e-11
#> [4,] 4.198121e-09 4.121742e-11
#> [5,] 4.033320e-09 5.122627e-11
#>