Simulate external exposure

Simulate external exposure

Usage

simulate_exposure(
  x,
  expos_mean = "mean",
  expos_sd = "sd",
  expos_label = "casn",
  n = 1000
)

Arguments

x

data frame or list of data frames containing exposure data.

expos_mean

column name of mean values.

expos_sd

column name of standard deviations.

expos_label

column name of labeling term, required if x has more than one row.

n

simulated sample size.

Value

list of matrices containing inhalation rates. Matrix columns are named using the values in the expos_label column for more than one data frame row. Columns are sorted to have consistent order across functions.

Examples

# Single data frame
x <- data.frame(mean = 1:3, sd = (1:3) / 10, casn = letters[1:3])
simulate_exposure(x, n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 1.2112277 2.039678 2.523414
#> [2,] 0.9643876 1.919919 2.838023
#> [3,] 0.8935536 2.123231 2.649162
#> [4,] 1.1077117 2.394831 3.167732
#> [5,] 1.1181576 2.376932 2.454196
#> 

# List of 2 data frames
y <- data.frame(mean = 4:6, sd = 0.1, casn = letters[1:3])
simulate_exposure(list(loc1 = x, loc2 = y), n = 5)
#> $loc1
#>              a        b        c
#> [1,] 1.0393344 2.039755 3.637933
#> [2,] 1.0042134 2.130107 2.557141
#> [3,] 1.1179664 2.068783 3.122367
#> [4,] 0.9743079 2.295506 3.418193
#> [5,] 0.8943664 2.014405 3.108083
#> 
#> $loc2
#>             a        b        c
#> [1,] 4.065455 5.029503 6.085291
#> [2,] 4.105216 5.126634 5.976566
#> [3,] 3.802044 4.886466 6.208669
#> [4,] 4.120839 4.886895 5.988908
#> [5,] 3.983072 5.010999 5.860715
#> 

# Input has custom column names
z <- data.frame(ave = 1:3, stdev = (1:3) / 10, chnm = letters[1:3])
simulate_exposure(z,
                  expos_mean = "ave",
                  expos_sd = "stdev",
                  expos_label = "chnm",
                  n = 5)
#> [[1]]
#>              a        b        c
#> [1,] 0.8857709 2.157293 3.434597
#> [2,] 1.1704609 1.884211 3.455458
#> [3,] 0.9919926 1.970915 2.884798
#> [4,] 0.9562719 2.105292 3.548138
#> [5,] 0.9880785 2.346716 2.834552
#>